unittests warning cleanup: fixed typechecks, rawtypes, redundant casts...
ScreenFinder fixed to work partially also in headless mode.
After this clean up only "internal proprietary API and may be removed in a future release" warnings remain fro make check. Please keep itw in this way :)

+++ tests/test-extensions/net/sourceforge/jnlp/awt/awtactions/KeyboardActions.java
@@ -98,9 +98,9 @@ public class KeyboardActions {
         int key;
         
         if( ('0' <= ch) && ('9' >= ch) ){
-            key = (int)(ch - '0') + KeyEvent.VK_0; 
+            key = (ch - '0') + KeyEvent.VK_0; 
         }else if( ( 'a' <= ch) && ('z' >= ch) ){
-            key = (int)(ch - 'a') + KeyEvent.VK_A;
+            key = (ch - 'a') + KeyEvent.VK_A;
         }else{
             key = KeyEvent.VK_SPACE;
         }
